fixed some warnings and removed some commented code

+++ samples/clients/cpp-nightfury/include/AudioPlayerEntry.h
@@ -5,7 +5,7 @@ namespace AudioPlayer{
     class AudioPlayerEntry{
         public:
             AudioPlayerEntry(unsigned char* pData, size_t pSize);
-            unsigned int m_size;
+            size_t m_size;
             unsigned char *m_data;
     };
 }
